Fantasy_BBall/requirements.txt
python-dotenv>=1.0.0
requests-oauthlib==2.0.0
pandas==2.2.2
# AI Integration
anthropic==0.39.0